pyproject.toml
[project]
name = "imageflow-backend"
version = "1.0.11"
requires-python = ">=3.10"
dependencies = [
  "pillow>=10.0.0",
  "piexif>=1.1.3",
  "exifread>=3.0.0",
  "reportlab>=4.0.0",
  "svglib>=1.5.1",
  "lxml>=5.0.0",
  "mozjpeg-lossless-optimization>=1.1.0",
  "imagequant>=1.0.0",
  "pyoxipng>=1.0.0",
  "pywebview>=5.3",
]

[dependency-groups]
build = [
  "pyinstaller>=6.13.0,<7.0.0",
]
